refactor: improve login screen design
