--- lib/home_page.dart
@@ -1,41 +1,0 @@
-import 'package:flutter/material.dart';
-
-class HomePage extends StatefulWidget {
-  const HomePage({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-
-  @override
-  Widget build(BuildContext context) {
-
-    return Scaffold(
-      appBar: AppBar(
-
-        title: Text(widget.title),
-      ),
-      body: Center(
-
-        child: Column(
-
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            ElevatedButton(
-              child: const Text("Sign In"),
-              onPressed: () {},
-            ),
-            ElevatedButton(
-              child: const Text("Sign Up"),
-              onPressed: () {},
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
