+++ lib/login_screen.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+class LoginScreen extends StatefulWidget {
+  const LoginScreen({Key? key, required this.title}) : super(key: key);
+
+  final String title;
+
+  @override
+  State<LoginScreen> createState() => _LoginScreenState();
+}
+
+class _LoginScreenState extends State<LoginScreen> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(widget.title),
+      ),
+      body: Center(
+        child: Padding(
+          padding: const EdgeInsets.all(24.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              Image.asset('assets/images/sloth.jpg'),
+              const SizedBox(
+                height: 16.0,
+              ),
+              SizedBox(
+                width: MediaQuery.of(context).size.width * 0.6,
+                child: ElevatedButton(
+                  child: const Text("Login"),
+                  onPressed: () {},
+                ),
+              ),
+              SizedBox(
+                width: MediaQuery.of(context).size.width * 0.6,
+                child: OutlinedButton(
+                  child: const Text("Crie uma conta"),
+                  onPressed: () {},
+                ),
+              )
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
